Trait sp_database::Database

pub trait Database<H: Clone + AsRef<[u8]>>: Send + Sync {
    fn commit(&self, transaction: Transaction<H>) -> Result<()>;

    fn get(&self, col: ColumnId, key: &[u8]) -> Option<Vec<u8>>;

    fn contains(&self, col: ColumnId, key: &[u8]) -> bool { ... }

    fn value_size(&self, col: ColumnId, key: &[u8]) -> Option<usize> { ... }

    fn with_get(&self, col: ColumnId, key: &[u8], f: &mut dyn FnMut(&[u8])) { ... }
}

Required methods

fn commit(&self, transaction: Transaction<H>) -> Result<()>

Commit the transaction to the database atomically. Any further calls to get or lookup will reflect the new state.

fn get(&self, col: ColumnId, key: &[u8]) -> Option<Vec<u8>>

Retrieve the value previously stored against key or None if key is not currently in the database.

Loading content...
Provided methods

fn contains(&self, col: ColumnId, key: &[u8]) -> bool

Check if the value exists in the database without retrieving it.

fn value_size(&self, col: ColumnId, key: &[u8]) -> Option<usize>

Check value size in the database possibly without retrieving it.

fn with_get(&self, col: ColumnId, key: &[u8], f: &mut dyn FnMut(&[u8]))

Call f with the value previously stored against key.

This may be faster than get since it doesn’t allocate. Use with_get helper function if you need f to return a value from f

Loading content...